mvp_sem_rag_openai_pkg2/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
jinja2==3.1.4
pdfplumber==0.11.4
pypdf==4.2.0
openai>=1.40.0
google-generativeai
